Finished the edge case of the Modal. Modularize the code
