+++ AltimetrikGameFinderApp/GameFinder/GameFinder.html
@@ -20,7 +20,6 @@
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.8.3/font/bootstrap-icons.css">
 
 </head>
-<body>
     <header>
         <div class="container-fluid ps-0 ">
            
@@ -50,8 +49,7 @@
                         <div class="form-group has-search">
                             <i class="bi-search form-control-feedback"></i>
                             <input id="searchBar" type="text" class="form-control" placeholder="Search games...">
-                            <div id="searchResults">
-                                
+                            <div id="searchResults">     
                             </div>
                         </div>
 
@@ -70,6 +68,7 @@
         </div>
         
     </header>
+    <div id="over">
     <div class="container-fluid ps-0">
         <!-- Begin container for the title, subtitle and the switch options-->
         <div class="container-fluid ps-4">
@@ -232,7 +231,8 @@
         <!-- End container for the sidebar and the cards-->                     
     </div>
         
-    <script src="scriptGameLoader.js"></script>
+    <script  type="module" src="scriptGameLoader.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.0-beta1/dist/js/bootstrap.bundle.min.js" integrity="sha384-pprn3073KE6tl6bjs2QrFaJGz5/SUsLqktiwsUTF55Jfv3qYSDhgCecCxMW52nD2" crossorigin="anonymous"></script>
 </body>
+</div>
 </html>
