+++ AltimetrikGameFinderApp/GameFinder/scriptGameLoader.js
@@ -1,441 +1,48 @@
-let gamesCards = ''
-let id = 1
-let pageCounter = 1
-let pageCounterForSearches = 1
-let lastGame
-let months = ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec']
-let response = []
-let searchResponse = []
-let likeButton 
-let Search = false
-let Queue = []
-let searchResultHtml = ''
-let resultsToShow = ''
-let resultsCounter = 4
-let gameSearch = ''
-let resp 
-let respSearch
-let savedYet = false
-let stopLoading = false
-let global 
-let consoleId
-let gameData = []
-let completeGameData = []
+import {getGameData , getGames , search} from './fetchModule.js'
+import {DisplayModal, DataRender} from './renderModule.js'
 
-
-const noBackground = "img/noBackground.svg"
+let overlay = document.getElementById('over')
+let icon = document.getElementById('icon')
+let user = document.getElementById('log-out')
 
 document.addEventListener('click', (e) => {
     if (e.target.closest('.card')){ 
         let cardToShow = e.target.innerHTML
         DisplayModal(cardToShow)
+    }
+    else{
+        overlay.classList.remove('overlay')
+        icon.classList.remove('overlay')
+        user.classList.remove('overlay')
     }  
 })
 
-/* begin getGames -> calls the api */
-const getGames = async() => {
-    try {
-        if (Search){
-            if (gameSearch == "Playstation" || gameSearch === "PC " || gameSearch === "Xbox" || gameSearch === "Nintendo"){ // it is a console
-                switch (gameSearch) {
-                    case ("Playstation"):
-                        consoleId = 2
-                    case ("PC "):
-                        consoleId = 1   
-                    case ("Xbox"):
-                        consoleId = 3
-                    case ("Nintendo"):
-                        consoleId = 7
-                        break;         
-                }
-                respSearch = await fetch(`https://api.rawg.io/api/games?key=6f2b18a8c113418299c8b7a82d268af9&parent_platforms=${consoleId}&page=${pageCounterForSearches}`)
-            } else {
-                respSearch = await fetch(`https://api.rawg.io/api/games?key=6f2b18a8c113418299c8b7a82d268af9&search=${gameSearch}&page=${pageCounterForSearches}`)
-            }
- 
-           
-            searchResponse = await respSearch.json()
-            if (respSearch.ok == true){
-                if (searchResponse.next == null) stopLoading = true;
-                getData(searchResponse.results)
-                pageCounterForSearches++                
-            }  else throw e       
-        } else {
-            resp = await fetch(`https://api.rawg.io/api/games?key=6f2b18a8c113418299c8b7a82d268af9&page=${pageCounter}`)
-            response = await resp.json()
-            if (resp.ok == true){
-                getData(response.results)
-                getGameData(response.results)
-            }
-                
-            else throw e
-             
-    }} catch (e) {
-        console.log('Sorry, something went wrong.')
-        console.log(e)
-    }
-}
-/* end getGames*/
-
-/* begin Observer -> it is the one in charge to ask for more games to the api*/
-let observer = new IntersectionObserver((gameObserved) => {
-    
-	    gameObserved.forEach(gameObserved => {
-		if(gameObserved.isIntersecting && !stopLoading){
-            if (!Search) pageCounter++;
-			getGames();
-		}
-	});
-}, {
-	rootMargin: '0px 0px 600px 0px',
-	threshold: 1.0
-}) 
-/* end Observer */
-
-let cardsx
-
-/* begin getData -> obtains all the data of the games */ 
-const getData = response =>{
-    response.forEach(game => {
-      
-       gamesCards += `
-        
-        <div class = "col-lg-4 d-block-lg col-md-6 py-2">
-            <div id= "card_${id}" class="card">
-                <input type ="image" class="like" value="false" src ="img/Heart.png">
-                <img class="card-img" id = "gameFrontPage" src="${game.background_image || noBackground}" alt="gameFrontPage">
-                <div class="card-body overflow-auto">
-                    <div class = "container g-0">
-                        <div class= "row">
-                            <div class= "col-10">
-                                <h3 class="card-title">${game.name}</h3>
-                            </div>
-                                
-                            <div class="col-2 pt-1">
-                                <p id ="gamePosition" > #${id} </p>
-                            </div>
-                        </div>
-                    </div>
-                        <div class="container g-0">
-                            <div class="row">
-                                <div class = "col-lg-4 col-5">
-                                    <p class="card-text date">Release date:</p>   
-                                </div>
-                                <div class = "col-lg-4 col-7" >
-                                    <p id="date" class= "card-text"> ${obtainDate(game.released)} </p>
-                                </div>
-                                <div class ="col-lg-4 ps-0 col-6">
-                                    <div id="platformIcons" class ="container">
-                                        <div class = "row platforms">
-                                            ${getPlat(game.parent_platforms)}
-                                        </div>
-                                    </div>
-                                </div>
-                            </div>
-
-                            <div id ="genresDiv" class="row pt-1">
-                                <div class = "col-lg-3 col-4">
-                                    <p class="card-text genres">Genres:</p>
-                                </div>
-                                <div class = "col-lg-9 col-8">
-                                    <p  class= "card-text "> ${ObtainInfo(game.genres, false)} </p>
-                                </div>
-                            </div>    
-                            
-                        </div>
-                    </div>
-        
-            </div>
-        </div>
-        `
-        
-        id++ 
-        
-        if (Search && savedYet){
-            savedYet = false 
-            if (Queue.length < 1)
-                Queue.push(gamesCards)
-            else{
-                Queue[1] = Queue[0]
-                Queue[0] = gamesCards 
-            }
-        }
-
-    });
-    
-    /* Last search funcionality*/
-    /* end last search func */
-    if (response.length > 0 ) {
-
-        document.getElementById('cards').innerHTML = gamesCards
-        
-        
-        if ((lastGame) && (id != 746582)){ // If id = 746582 it is the last game
-            observer.unobserve(lastGame)
-        }
-
-        const gamesDisplayed = document.querySelectorAll('.card')
-        lastGame = gamesDisplayed[gamesDisplayed.length - 1]
-        observer.observe(lastGame)
-    } 
-}
-/* end getData*/
-
-/* begin ObtainInfo -> obtains info of the games */
-function ObtainInfo(info, flag){
-    let infoCards = ''
-    if (info){
-        for (let index = 0; index < info.length; index++) {
-            if (!flag) 
-                infoCards += info[index].name
-            else 
-                infoCards += info[index].platform.name;
-            if (index + 1 != info.length) {  // If there are more than one genre.
-                infoCards += ', '
-            }
-        }
-    } else infoCards = 'The publisher did not specified'
-   return infoCards
-}
-/* end ObtainInfo */
-
-
-/* begin getPlat -> obtains the platforms of the games */
-const getPlat  = platforms =>{
-    let platformCards = ''
-    if (platforms){
-        for (let index = 0; index < platforms.length; index++) {
-            const element = platforms[index];
-                
-                switch  (element.platform.id){
-                    case (2): // Play
-                        platformCards += `
-                            <div class = "col-4">
-                                <svg  id ="PS" class="icons" width="24" height="20" viewBox="0 0 24 20" fill="#FFFFFF" xmlns="http://www.w3.org/2000/svg">
-                                <path d="M9.55164 9.53674e-06L9.55164 18.2774L13.3604 19.5885L13.3604 4.26319C13.3604 3.54103 13.6567 3.06155 14.132 3.22645C14.7534 3.41372 14.8742 4.07895 14.8742 4.79307L14.8742 10.9138C17.2449 12.1606 19.1114 10.9131 19.1114 7.62345C19.1114 4.26188 18.0171 2.76435 14.7972 1.56033C13.5271 1.10084 11.1734 0.325457 9.55164 9.53674e-06Z" fill="##FFFFFF"/>
-                                <path d="M14.3503 16.9129L20.1076 14.3203C20.7589 14.0148 20.8585 13.5998 20.3314 13.3817C19.796 13.1596 18.8404 13.2232 18.1821 13.5222L14.3503 15.2325V12.5036L14.5698 12.412C14.5698 12.412 15.6789 11.9148 17.239 11.7008C18.796 11.4848 20.7055 11.7289 22.2063 12.4446C23.8969 13.1241 24.0864 14.1153 23.6587 14.8051C23.2247 15.4877 22.1726 15.9815 22.1726 15.9815L14.3503 19.5367" fill="##FFFFFF"/>
-                                <path d="M1.7099 17.2146C-0.0868735 16.6444 -0.386499 15.4395 0.433179 14.7435C1.18926 14.1079 2.47691 13.6295 2.47691 13.6295L7.79974 11.4675V13.9281L3.97289 15.4908C3.29503 15.767 3.1934 16.1557 3.73939 16.359C4.29514 16.5706 5.28141 16.5137 5.95966 16.2286L7.79974 15.4728V17.6695C7.68114 17.6921 7.54927 17.715 7.42892 17.7384C5.5943 18.0852 3.63971 17.9428 1.7099 17.2146Z" fill="##FFFFFF"/>
-                                <path fill-rule="evenodd" clip-rule="evenodd" d="M23.7661 19.4589C23.6149 19.6145 23.4164 19.7024 23.2027 19.7024C22.989 19.7024 22.784 19.6145 22.6326 19.4589C22.4831 19.3005 22.4004 19.0936 22.4004 18.8706C22.4004 18.4089 22.7591 18.0357 23.2027 18.0357C23.4164 18.0357 23.6149 18.1208 23.7661 18.2798C23.9156 18.4352 24 18.6456 24 18.8706C24 19.0936 23.9156 19.3005 23.7661 19.4589ZM22.5352 18.8707C22.5352 18.6808 22.6033 18.5067 22.7279 18.3776C22.8555 18.2458 23.0258 18.1747 23.2027 18.1747C23.3798 18.1747 23.5458 18.2458 23.6703 18.3776C23.7959 18.5067 23.8638 18.6808 23.8638 18.8707C23.8638 19.2511 23.567 19.5599 23.2027 19.5599C23.0258 19.5599 22.8555 19.4896 22.7279 19.3594C22.6033 19.2281 22.5352 19.0558 22.5352 18.8707ZM23.5677 19.2169C23.5748 19.2384 23.5835 19.2511 23.5958 19.2548L23.607 19.2614V19.3143H23.4334L23.4302 19.3036L23.4184 19.2717C23.4164 19.2548 23.4141 19.2328 23.4117 19.1957L23.404 19.0508C23.402 18.9993 23.3859 18.9694 23.3561 18.9496C23.334 18.9419 23.3039 18.9359 23.259 18.9359H23.018V19.3143H22.8599V18.3849H23.2745C23.3421 18.3849 23.399 18.3975 23.4426 18.4167C23.53 18.4596 23.5748 18.5368 23.5748 18.6455C23.5748 18.6988 23.5621 18.7487 23.5402 18.7856C23.5212 18.8116 23.4988 18.8353 23.4744 18.8585L23.4809 18.8633C23.4974 18.8754 23.5138 18.8874 23.5235 18.9051C23.5456 18.9305 23.5556 18.9731 23.5574 19.028L23.5614 19.1463C23.5621 19.1766 23.5644 19.2002 23.5677 19.2169ZM23.3807 18.7599C23.4063 18.7427 23.4184 18.7085 23.4184 18.6561C23.4184 18.6009 23.4001 18.5641 23.3642 18.5458C23.3421 18.5368 23.3146 18.5303 23.2779 18.5303H23.018V18.7914H23.2635C23.3123 18.7914 23.3511 18.7809 23.3807 18.7599Z" fill="#FFFFFF"/>
-                                </svg>
-                            </div>
-                            `  
-                        break 
-                    case (3): // Xbox
-                        
-                        platformCards += `
-                            <div class = "col-4">
-                                <svg  id="Xbox" class="icons" width="20" height="20" viewBox="0 0 20 20" fill="#FFFFFF" xmlns="http://www.w3.org/2000/svg">
-                                <path class="icons" fill-rule="evenodd" clip-rule="evenodd" d="M10 0C11.9286 0 13.5238 0.616246 14.9762 1.65266C15 1.65266 15 1.68067 15 1.70868C15 1.73669 14.9762 1.73669 14.9524 1.73669C13.0952 1.2605 10.2857 3.13725 10.0238 3.33333H10H9.97619C9.71429 3.13725 6.90476 1.2605 5.04762 1.73669C5.02381 1.73669 5 1.73669 5 1.70868C5 1.68067 5 1.65266 5.02381 1.65266C6.47619 0.616246 8.07143 0 10 0ZM16.3903 17.5988C17.8903 16.0464 12.9308 10.5648 10.0035 8.33333C10.0035 8.33333 9.97935 8.33333 9.97935 8.35759C7.07626 10.5648 2.09261 16.0464 3.61674 17.5988C5.31021 19.1026 7.56011 20 10.0035 20C12.447 20 14.6727 19.1026 16.3903 17.5988ZM2.73973 3.38078C2.72831 3.38078 2.7226 3.38705 2.7169 3.39332C2.71119 3.39959 2.70548 3.40585 2.69406 3.40585C1.0274 5.2358 0 7.76763 0 10.5501C0 12.8313 0.707763 14.9621 1.87215 16.6416C1.87215 16.6667 1.89498 16.6667 1.91781 16.6667C1.94064 16.6667 1.94064 16.6416 1.91781 16.6165C1.21005 14.2351 4.79452 8.4946 6.64384 6.0881L6.66667 6.06303C6.66667 6.03796 6.66667 6.03796 6.64384 6.03796C3.83562 2.9797 2.89954 3.30558 2.73973 3.38078ZM13.3333 6.05268L13.3562 6.02759C16.1644 2.99144 17.1005 3.31764 17.2374 3.36782C17.2469 3.36782 17.2525 3.36782 17.2574 3.36962C17.2642 3.37215 17.2698 3.37825 17.2831 3.39291C18.9726 5.22464 20 7.75895 20 10.5442C20 12.8276 19.2922 14.9604 18.1279 16.6416C18.1279 16.6667 18.105 16.6667 18.0822 16.6667V16.6165C18.7671 14.2327 15.2055 8.48662 13.3562 6.07777C13.3333 6.07777 13.3333 6.05268 13.3333 6.05268Z" fill="#FFFFFF"/>
-                                </svg>
-                            </div>
-                        `
-                    break
-                    case (7): // Nintendo
-
-                        platformCards += `
-                            <div class = "col-4">
-                                <svg id="Nintendo" class="icons" width="24" height="24" viewBox="0 0 24 24" fill="#FFFFFF" xmlns="http://www.w3.org/2000/svg">
-                                <path fill-rule="evenodd" clip-rule="evenodd" d="M17.8608 24H14.1696C14.0755 24 14.0001 23.9245 14.0001 23.8302V0.150943C14.0001 0.0754717 14.0566 0 14.1508 0H17.8608C21.2506 0 24.0001 2.75472 24.0001 6.15094V17.8491C24.0001 21.2453 21.2506 24 17.8608 24ZM21.1564 13.2076C21.1564 11.8679 20.0641 10.7736 18.727 10.7736C17.3899 10.7736 16.3165 11.8679 16.2976 13.2076C16.2976 14.5472 17.3899 15.6415 18.727 15.6415C20.0641 15.6415 21.1564 14.5472 21.1564 13.2076Z" fill="##FFFFFF"/>
-                                <path d="M3.99988 8C3.99988 9.1 4.89988 10 5.99988 10C7.09988 10 7.99988 9.1 7.99988 8C7.99988 6.9 7.09988 6 5.99988 6C4.88321 6 3.99988 6.88333 3.99988 8Z" fill="##FFFFFF"/>
-                                <path fill-rule="evenodd" clip-rule="evenodd" d="M6.38173 0H11.8238C11.9217 0 12 0.0754717 12 0.169811V23.8302C12 23.9245 11.9217 24 11.8238 24H6.38173C2.85807 24 0 21.2453 0 17.8491V6.15094C0 2.75472 2.85807 0 6.38173 0ZM6.38169 22.0755H10.0032V1.92458H6.38169C5.20714 1.92458 4.11089 2.37741 3.2887 3.16986C2.44694 3.96232 1.99669 5.01892 1.99669 6.15099V17.8491C1.99669 18.9812 2.46652 20.0378 3.2887 20.8302C4.11089 21.6416 5.20714 22.0755 6.38169 22.0755Z" fill="##FFFFFF"/>
-                                </svg> 
-                            </div>  
-                        `  
-                        break
-                    case (1): // PC
-
-                        platformCards += `
-                            <div class = "col-4">
-                                <svg id ="PC" class="icons" width="20" height="20" viewBox="0 0 20 20" fill="#FFFFFF" xmlns="http://www.w3.org/2000/svg">
-                                <path fill-rule="evenodd" clip-rule="evenodd" d="M20 9.16667H9.16667V1.53647L20 0V9.16667ZM8.33333 1.66667V9.16667H0V2.77865L8.33333 1.66667ZM8.33333 10H0V17.0992L8.33333 18.3333V10ZM9.16667 18.3262V10H20V20L9.16667 18.3262Z" fill="##FFFFFF"/>
-                                </svg>
-                            </div> 
-                        ` 
-                        break  
-                    default:
-                        break            
-                }
-        }
-    } else {
-        platformCards = 'There are not platform associated to this game'
-    }   
-    return platformCards 
-   
-}
-/* end getPlat  */
-
-
-const getPlatformsModal = platforms  => {
-    let platformModal = ''
-    if (platforms){
-        for (let index = 0; index < platforms.length; index++) {
-            const element = platforms[index];
-                
-                switch  (element.platform.id){
-                    case (2): // Play
-                        platformModal += `
-                            
-                                <svg  id ="PS" class="iconsModal" width="32" height="32" viewBox="0 0 24 20" fill="#FFFFFF" xmlns="http://www.w3.org/2000/svg">
-                                <path d="M9.55164 9.53674e-06L9.55164 18.2774L13.3604 19.5885L13.3604 4.26319C13.3604 3.54103 13.6567 3.06155 14.132 3.22645C14.7534 3.41372 14.8742 4.07895 14.8742 4.79307L14.8742 10.9138C17.2449 12.1606 19.1114 10.9131 19.1114 7.62345C19.1114 4.26188 18.0171 2.76435 14.7972 1.56033C13.5271 1.10084 11.1734 0.325457 9.55164 9.53674e-06Z" fill="##FFFFFF"/>
-                                <path d="M14.3503 16.9129L20.1076 14.3203C20.7589 14.0148 20.8585 13.5998 20.3314 13.3817C19.796 13.1596 18.8404 13.2232 18.1821 13.5222L14.3503 15.2325V12.5036L14.5698 12.412C14.5698 12.412 15.6789 11.9148 17.239 11.7008C18.796 11.4848 20.7055 11.7289 22.2063 12.4446C23.8969 13.1241 24.0864 14.1153 23.6587 14.8051C23.2247 15.4877 22.1726 15.9815 22.1726 15.9815L14.3503 19.5367" fill="##FFFFFF"/>
-                                <path d="M1.7099 17.2146C-0.0868735 16.6444 -0.386499 15.4395 0.433179 14.7435C1.18926 14.1079 2.47691 13.6295 2.47691 13.6295L7.79974 11.4675V13.9281L3.97289 15.4908C3.29503 15.767 3.1934 16.1557 3.73939 16.359C4.29514 16.5706 5.28141 16.5137 5.95966 16.2286L7.79974 15.4728V17.6695C7.68114 17.6921 7.54927 17.715 7.42892 17.7384C5.5943 18.0852 3.63971 17.9428 1.7099 17.2146Z" fill="##FFFFFF"/>
-                                <path fill-rule="evenodd" clip-rule="evenodd" d="M23.7661 19.4589C23.6149 19.6145 23.4164 19.7024 23.2027 19.7024C22.989 19.7024 22.784 19.6145 22.6326 19.4589C22.4831 19.3005 22.4004 19.0936 22.4004 18.8706C22.4004 18.4089 22.7591 18.0357 23.2027 18.0357C23.4164 18.0357 23.6149 18.1208 23.7661 18.2798C23.9156 18.4352 24 18.6456 24 18.8706C24 19.0936 23.9156 19.3005 23.7661 19.4589ZM22.5352 18.8707C22.5352 18.6808 22.6033 18.5067 22.7279 18.3776C22.8555 18.2458 23.0258 18.1747 23.2027 18.1747C23.3798 18.1747 23.5458 18.2458 23.6703 18.3776C23.7959 18.5067 23.8638 18.6808 23.8638 18.8707C23.8638 19.2511 23.567 19.5599 23.2027 19.5599C23.0258 19.5599 22.8555 19.4896 22.7279 19.3594C22.6033 19.2281 22.5352 19.0558 22.5352 18.8707ZM23.5677 19.2169C23.5748 19.2384 23.5835 19.2511 23.5958 19.2548L23.607 19.2614V19.3143H23.4334L23.4302 19.3036L23.4184 19.2717C23.4164 19.2548 23.4141 19.2328 23.4117 19.1957L23.404 19.0508C23.402 18.9993 23.3859 18.9694 23.3561 18.9496C23.334 18.9419 23.3039 18.9359 23.259 18.9359H23.018V19.3143H22.8599V18.3849H23.2745C23.3421 18.3849 23.399 18.3975 23.4426 18.4167C23.53 18.4596 23.5748 18.5368 23.5748 18.6455C23.5748 18.6988 23.5621 18.7487 23.5402 18.7856C23.5212 18.8116 23.4988 18.8353 23.4744 18.8585L23.4809 18.8633C23.4974 18.8754 23.5138 18.8874 23.5235 18.9051C23.5456 18.9305 23.5556 18.9731 23.5574 19.028L23.5614 19.1463C23.5621 19.1766 23.5644 19.2002 23.5677 19.2169ZM23.3807 18.7599C23.4063 18.7427 23.4184 18.7085 23.4184 18.6561C23.4184 18.6009 23.4001 18.5641 23.3642 18.5458C23.3421 18.5368 23.3146 18.5303 23.2779 18.5303H23.018V18.7914H23.2635C23.3123 18.7914 23.3511 18.7809 23.3807 18.7599Z" fill="#FFFFFF"/>
-                                </svg>
-                           
-                            `  
-                        break 
-                    case (3): // Xbox
-                        
-                        platformModal += `
-                            
-                                <svg  id="Xbox" class="iconsModal" width="32" height="32" viewBox="0 0 20 20" fill="#FFFFFF" xmlns="http://www.w3.org/2000/svg">
-                                <path class="icons" fill-rule="evenodd" clip-rule="evenodd" d="M10 0C11.9286 0 13.5238 0.616246 14.9762 1.65266C15 1.65266 15 1.68067 15 1.70868C15 1.73669 14.9762 1.73669 14.9524 1.73669C13.0952 1.2605 10.2857 3.13725 10.0238 3.33333H10H9.97619C9.71429 3.13725 6.90476 1.2605 5.04762 1.73669C5.02381 1.73669 5 1.73669 5 1.70868C5 1.68067 5 1.65266 5.02381 1.65266C6.47619 0.616246 8.07143 0 10 0ZM16.3903 17.5988C17.8903 16.0464 12.9308 10.5648 10.0035 8.33333C10.0035 8.33333 9.97935 8.33333 9.97935 8.35759C7.07626 10.5648 2.09261 16.0464 3.61674 17.5988C5.31021 19.1026 7.56011 20 10.0035 20C12.447 20 14.6727 19.1026 16.3903 17.5988ZM2.73973 3.38078C2.72831 3.38078 2.7226 3.38705 2.7169 3.39332C2.71119 3.39959 2.70548 3.40585 2.69406 3.40585C1.0274 5.2358 0 7.76763 0 10.5501C0 12.8313 0.707763 14.9621 1.87215 16.6416C1.87215 16.6667 1.89498 16.6667 1.91781 16.6667C1.94064 16.6667 1.94064 16.6416 1.91781 16.6165C1.21005 14.2351 4.79452 8.4946 6.64384 6.0881L6.66667 6.06303C6.66667 6.03796 6.66667 6.03796 6.64384 6.03796C3.83562 2.9797 2.89954 3.30558 2.73973 3.38078ZM13.3333 6.05268L13.3562 6.02759C16.1644 2.99144 17.1005 3.31764 17.2374 3.36782C17.2469 3.36782 17.2525 3.36782 17.2574 3.36962C17.2642 3.37215 17.2698 3.37825 17.2831 3.39291C18.9726 5.22464 20 7.75895 20 10.5442C20 12.8276 19.2922 14.9604 18.1279 16.6416C18.1279 16.6667 18.105 16.6667 18.0822 16.6667V16.6165C18.7671 14.2327 15.2055 8.48662 13.3562 6.07777C13.3333 6.07777 13.3333 6.05268 13.3333 6.05268Z" fill="#FFFFFF"/>
-                                </svg>
-                          
-                        `
-                    break
-                    case (7): // Nintendo
-
-                        platformModal += `
-                            
-                                <svg id="Nintendo" class="iconsModal" width="32" height="32" viewBox="0 0 24 24" fill="#FFFFFF" xmlns="http://www.w3.org/2000/svg">
-                                <path fill-rule="evenodd" clip-rule="evenodd" d="M17.8608 24H14.1696C14.0755 24 14.0001 23.9245 14.0001 23.8302V0.150943C14.0001 0.0754717 14.0566 0 14.1508 0H17.8608C21.2506 0 24.0001 2.75472 24.0001 6.15094V17.8491C24.0001 21.2453 21.2506 24 17.8608 24ZM21.1564 13.2076C21.1564 11.8679 20.0641 10.7736 18.727 10.7736C17.3899 10.7736 16.3165 11.8679 16.2976 13.2076C16.2976 14.5472 17.3899 15.6415 18.727 15.6415C20.0641 15.6415 21.1564 14.5472 21.1564 13.2076Z" fill="##FFFFFF"/>
-                                <path d="M3.99988 8C3.99988 9.1 4.89988 10 5.99988 10C7.09988 10 7.99988 9.1 7.99988 8C7.99988 6.9 7.09988 6 5.99988 6C4.88321 6 3.99988 6.88333 3.99988 8Z" fill="##FFFFFF"/>
-                                <path fill-rule="evenodd" clip-rule="evenodd" d="M6.38173 0H11.8238C11.9217 0 12 0.0754717 12 0.169811V23.8302C12 23.9245 11.9217 24 11.8238 24H6.38173C2.85807 24 0 21.2453 0 17.8491V6.15094C0 2.75472 2.85807 0 6.38173 0ZM6.38169 22.0755H10.0032V1.92458H6.38169C5.20714 1.92458 4.11089 2.37741 3.2887 3.16986C2.44694 3.96232 1.99669 5.01892 1.99669 6.15099V17.8491C1.99669 18.9812 2.46652 20.0378 3.2887 20.8302C4.11089 21.6416 5.20714 22.0755 6.38169 22.0755Z" fill="##FFFFFF"/>
-                                </svg> 
-                              
-                        `  
-                        break
-                    case (1): // PC
-
-                        platformModal += `
-                           
-                                <svg id ="PC" class="iconsModal" width="32" height="32" viewBox="0 0 20 20" fill="#FFFFFF" xmlns="http://www.w3.org/2000/svg">
-                                <path fill-rule="evenodd" clip-rule="evenodd" d="M20 9.16667H9.16667V1.53647L20 0V9.16667ZM8.33333 1.66667V9.16667H0V2.77865L8.33333 1.66667ZM8.33333 10H0V17.0992L8.33333 18.3333V10ZM9.16667 18.3262V10H20V20L9.16667 18.3262Z" fill="##FFFFFF"/>
-                                </svg>
-                            
-                        ` 
-                        break  
-                    default:
-                        break            
-                }
-        }
-    } else {
-        platformModal = 'There are not platform associated to this game'
-    }   
-    return platformModal
-} 
-
-/*begin obtainDate -> translate the date from ISO to the way it is asked*/
-let obtainDate = date => {
-    return months[new Date(date).getMonth()] + " " +
-    new Date(date).getUTCDate() + ", " +
-    new Date(date).getFullYear()
-}
-/* end obtainDate */ 
-
-/* begin searchBar */
-
-const debounce = (fn, delay=500) => {
-    let timeoutId
-
-    return (...args) => {
-       
-        if (timeoutId) {
-            clearTimeout(timeoutId)
-        }
-        timeoutId = setTimeout(() => {
-            fn.apply(null, args)
-        }, delay)
-    };
-};
-
-const searchTermElem = document.getElementById('searchBar')
+let searchTermElem = document.getElementById('searchBar')
 searchTermElem.select()
 
 searchTermElem.addEventListener('input', function (event) {
     search(event.target.value)
-    
-});
-
-    const showResults = searchTerm => {
-        document.getElementById('searchResults').innerHTML = ''
-        resultsCounter = searchResponse.results.length
-        if (resultsCounter >= 1){
-            for (let i = 0; i < resultsCounter ; i++) {      
-                searchResultHtml += `                
-                        <li> ${searchTerm.results[i].name} </li>
-                    ` 
-
-            }          
-            document.getElementById('searchResults').innerHTML = searchResultHtml
-            searchResultHtml = '' // To make it empty again
-        } else {
-            document.getElementById('searchResults').innerHTML = ''
-            resultsToShow = `
-                <li> The game that you are looking for does not exist in our database.</li>
-            ` 
-            document.getElementById('searchResults').innerHTML =  resultsToShow
-            resultsToShow = '' // To make it empty again
-        }
-    }
-
-
-    async function lookForResults(){
-        id = 1
-        gamesCards = ''
-        Search = true // boolean variable to dont make more fetch
-        savedYet = true // boolean variable just to save the first game
-        stopLoading = false
-        await getGames()
-        showResults(searchResponse) 
-    }
-
-    const search = debounce(async (searchTerm) => {
-
-        // if the search term is removed, reset the search result
-        if (!searchTerm || searchTerm.length < 3){
-            document.getElementById('searchResults').innerHTML = ''
-            return
-        } 
-    
-        try {
-            document.getElementById('searchResults').innerHTML = ''
-            gameSearch = searchTerm 
-            pageCounterForSearches = 1 
-            lookForResults()       
-        } catch (error) {
-            console.log(error);
-        }
-});
-/* end searchBar */
-
-
-
-/* Begin lastSearches */
-
-
-const lastSearches = document.getElementById('LastSearchButton')
-let LastGameShow = ''
-let gamesSearched = ''
-lastSearches.addEventListener('click', () =>{
-        document.getElementById('searchResults').innerHTML = '';
-        if (Queue.length == 0){
-            gamesSearched += `
-                <span class = "emptySearch"> You have not done any searches recently.</span>
-            `         
-        } else {
-            for (let index = 0; index < Queue.length ; index++) {
-                gamesSearched += Queue[index] 
-            }
-        }
-        document.getElementById('cards').innerHTML = gamesSearched
-        gamesSearched = ''
+    overlay.classList.add('overlay')
+    icon.classList.add('overlay')
+    user.classList.add('overlay')
 })
 
-/* end lastSearches */
-
-/* begin lightMode */
+const lastSearches = document.getElementById('LastSearchButton')
+let gamesSearched = ''
+lastSearches.addEventListener('click', () =>{
+    document.getElementById('searchResults').innerHTML = '';
+    if (DataRender.Queue.length == 0){
+        gamesSearched += `
+            <span class = "emptySearch"> You have not done any searches recently.</span>
+        `         
+    } else {
+        for (let index = 0; index < DataRender.Queue.length ; index++) {
+            gamesSearched += DataRender.Queue[index] 
+        }
+    }
+    document.getElementById('cards').innerHTML = gamesSearched
+    gamesSearched = ''
+})
 
 const switchModes = document.getElementById('flexSwitchCheckChecked')
 
@@ -449,159 +56,6 @@ switchModesMobile.addEventListener('click', () => {
     document.body.classList.toggle("lightmode")
 })
 
-/* end lightMode */
-
-
-/* begin search mobile */
-
-const searchMobile = document.getElementById('mobSearch')
-
-searchMobile.addEventListener('click', () => {
-    /*Work in progress */
-})
-
-/* end search mobile */
-let listOfGames = []
-function getGameData(gameData) {
-    for (let index = 0; index < gameData.length; index++) {
-      const gameId = gameData[index].id 
-      const gameScreenshots = gameData[index].short_screenshots 
-      fetch(`https://api.rawg.io/api/games/${gameId}?key=6f2b18a8c113418299c8b7a82d268af9&`)
-        .then((res) => res.json())
-        .then((data) => {
-            fetch(`https://api.rawg.io/api/games/${gameId}/movies?key=6f2b18a8c113418299c8b7a82d268af9&`)
-                .then((resMovie) => resMovie.json())
-                .then((dataMovie) => {
-                    Object.assign(data, gameScreenshots)
-                    Object.assign(data, dataMovie)
-                    listOfGames.push(data)
-            })
-        })
-    }
-}
-
-const ModalShown = document.getElementById('ModalInjection')
-/* begin modal fuctionality*/
-let ModalToShow = null
-function DisplayModal(CardTitle){
-    let find = false
-    let movieOfGame 
-    if (ModalToShow !== null){
-        ModalToShow.innerHTML = ''
-        ModalShown.innerHTML = ''
-    }
-    ModalToShow = document.createElement('div')
-    let i = 0
-    while (!find){
-        if (listOfGames[i].name === CardTitle){
-           if (listOfGames[i].results[0] != null)
-                movieOfGame = listOfGames[i].results[0].data['480']
-            else
-                movieOfGame = noBackground 
-            ModalToShow.innerHTML =
-            `
-                <div class="modal fade" id="Modal" tabindex="-1" aria-labelledby="ModalLabel" aria-hidden="true">
-                <div class="modal-dialog modal-xl modal-dialog-centered">
-                        <div class="modal-content">                        
-                            <img class="modalBackground" src="${listOfGames[i].background_image}" alt="Background of the Game">
-                            <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-                            <div class="container-fluid">
-                                <div class="modal-body">
-                                    <div class="container-fluid bigContainer">
-                                        <div class = "row pt-5 mt-2">
-                                            <div class = "col-12 iconsM"> 
-                                                ${getPlatformsModal(listOfGames[i].parent_platforms)}
-                                            </div>
-                                        </div>
-                                        <h3>${listOfGames[i].name}</h3>
-                                    </div>
-                                    <div class="container-fluid">
-                                        <div class="row">
-                                        <div class="col-7">
-                                            <div class = "gameDescription">
-                                                ${listOfGames[i].description}
-                                            </div>
-                                        <div class="container-fluid generalContainer">
-                                            <div class="row">
-                                                <div class="col-5">
-                                                    <button type="button" class="btn btn-primary wishlistButton">Add to wishlist</button>
-                                                    <div class = "infoAboutTheGame">
-                                                        <p class="platform">Platforms</p>
-                                                        <a href="">${ObtainInfo(listOfGames[i].platforms,true)}</a>
-
-                                                        <p class="date">Release date</p>
-                                                        <p>${obtainDate(listOfGames[i].released, false)}</p>
-
-                                                        <p class="publisher">Publisher</p>
-                                                        <a href="">${ObtainInfo(listOfGames[i].publishers, false)}</a>
-
-                                                        <p class="web">Website</p>
-                                                        <a href="">${listOfGames[i].website}</a>
-                                                    </div>
-                                                </div>
-                                                <div class="col-5">
-                                                    <button type="button" class="btn btn-outline-primary purchaseButton">Purchase</button>
-                                                    <div class = "infoAboutTheGame">
-                                                        <p class="genres">Genre</p>
-                                                        <a href="">${ObtainInfo(listOfGames[i].genres, false)}</a>
-
-                                                        <p class="dev">Developer</p>
-                                                        <a href="">${ObtainInfo(listOfGames[i].developers, false)}</a>
-
-                                                        <p class="age">Age rating</p>
-                                                        <p>${ObtainInfo(listOfGames[i].esrb_rating, false)}</p>
-                                                    </div>
-
-                                                </div>
-                                                <div class="col-2">
-                                                </div>
-                                            </div>
-                                        </div>            
-                                        </div>                   
-                                        <div class="col-5">
-                                            <div class = "row">                           
-                                                <video class = "gameTrailer" controls width="640" height="480">
-                                                    <source src="${movieOfGame}"/>
-                                                </video>
-                                            </div>
-                                            <div class = "row">
-                                                <div class = "col-6">
-                                                    <img class = "scshot" src="${listOfGames[i][1].image}">
-                                                </div>
-                                                <div class = "col-6">
-                                                    <img class = "scshot" src="${listOfGames[i][2].image}">
-                                                </div>
-                                            </div>
-                                            <div class = "row">
-                                                <div class = "col-6">
-                                                    <img class = "scshot" src="${listOfGames[i][3].image}">
-                                                </div>
-                                                <div class = "col-6">
-                                                    <img class = "scshot" src="${listOfGames[i][4].image}">
-                                                </div>
-                                            </div>
-                                        </div>
-                                    </div>
-                                    </div>
-                                    </div>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-                `
-                find = true 
-        }
-        i++
-    }
-    ModalShown.append(ModalToShow)
-    const myModal = new bootstrap.Modal(document.getElementById('Modal'))
-    myModal.show()
-    find = false
-}
-   
-
-/* end modal */ 
 getGames()
